pyproject.toml
[project]
name = "madargani-notemind"
version = "0.1.0"
description = "Transform your notes into a smart searchable knowledge base."
authors = [
    { name = "madargani", email = "mad3913@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
dependencies = [
    "chromadb (>=1.0.15,<2.0.0)",
    "pymupdf4llm (>=0.0.27,<0.0.28)",
    "streamlit (>=1.47.1,<2.0.0)",
    "appdirs (>=1.4.4,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-watch = "^4.2.0"
